Introducing 3 hexdump commands
`hexdump -C` is used in verbose mode to see what is going on. Unfortunately, it's not possible to get more than 16 columns
`hexdump -ve '16/1 "%02x " " \n"'` is used to analyse the reply given by the server. There is no offset and no human readable character output
`hexdump -ve '1/1 "%.2x"'` is used to generate a pure hexadecimal stream of data