Update OpenSSL reqs - LOGJAM checks need 1.0.2+
To effectively analyze the LOGJAM risks, and to display the bitstrength on the DH/ECDH negotiated ciphers, OpenSSL 1.0.2+ is needed.  With anything under 1.0.2 (and greater than 1.0.0), the bitstrengths are not displayed as OpenSSL is 'too old' (as referred to in the script itself when 1.0.2 is newer than what's available).

I suggest that we keep a note that >= 1.0.2 is needed for LOGJAM checks.