Increase HEADER_MAXSLEEP value to 10
There is a case where using this on a network with some (slight) instability or slowness can sometimes fail to detect HTTP headers.  Currently, in my testing here, I've seen this range anywhere from 5.7 seconds to 7.8 seconds for a response from getting the headers, so maybe we should consider increasing the time slightly for the header max sleep value.  10 seconds is arbitrary, but still 'short'.

(Feel free to discuss or decline, but I'm hoping we can incorporate this to adapt for slightly slower networks)